development

+++ app/Http/Controllers/BlogController.php
@@ -4,9 +4,12 @@ namespace App\Http\Controllers;
 
 use Inertia\Inertia;
 use App\Models\Blog;
+use App\Models\User;
 use Illuminate\Support\Facades\Request;
 use Illuminate\Support\Facades\Redirect;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Auth;
+
 
 class BlogController extends Controller
 {
@@ -20,6 +23,9 @@ class BlogController extends Controller
                         'title' => $blog->title,
                         'content' => $blog->content,
                         'image' => asset('storage/' . $blog->image),
+                        'user_id' => $blog->user_id,
+                        'user_name' => User::find($blog->user_id)->name,
+                        'logged_in' => Auth::id(),
                     ];
                 })
             ]
@@ -39,6 +45,7 @@ class BlogController extends Controller
             'title' => Request::input('title'),
             'content' => Request::input('content'),
             'image' => $image,
+            'user_id' => auth()->user()->id,
         ]);
 
         return Redirect::route('blog.index');
@@ -48,7 +55,8 @@ class BlogController extends Controller
     {
         return Inertia::render('Blogs/Edit', [
             'blog' => $blog,
-            'image' => asset('storage/' . $blog->image)
+            'image' => asset('storage/' . $blog->image),
+            'logged_in' => Auth::id(),
         ]);
     }
 
@@ -73,6 +81,10 @@ class BlogController extends Controller
 
     public function delete(Blog $blog)
     {
+        if($blog->user_id != Auth::id()){
+            return Redirect::route('blog.index');
+        }
+        
         Storage::delete('public/'. $blog->image);
         $blog->delete();
